Fixed bug when you can turn 180 degrees.
Bug: If you change direction two times in one tick, the second of which will be opposite of the initial direction,
the snake will turn 180 degrees in the next tick, which will lead to defeat.
Now new variable 'nextDirection' stores the desired direction change, and the current direction will be changed only in move() method.
When a player presses arrow keys, it changes only 'nextDirection' variable, and it does it only in that case,
when the arrow key is not opposite of the current direction.