Two direction changes in a tick are valid now.
Previously, due to fixing a '180-degree turn bug', a player was not allowed to make direction changes in one tick.
Now, when a player changes direction 2 times in one tick - it registers as a 'sharp turn'.
The game will automatically execute the first direction change in the current tick, and the second direction change will be scheduled for the next tick.