Added score, win screen, fixed 'endless loop' bug.
Bug: Previously, when the player ate the last possible apple (after which the win screen should appear), the game froze
because an endless loop occurs in the generateApple() method. There was not any free tile for apple to generate.
Now, the game tracks if the player has eaten the last apple, and in that case, shows a win screen.
Win screen contains the final score.
Now, the score is displayed in the top right corner during the game.